Rename dsn module to network
